==> notebooks/01_merge_titles_github.ipynb <==
# GitHub: Merge titles

- <a href="https://github.com/bhargaviparanjape/clickbait">[bhargaviparanjape] clickbait</a>
- <a href="https://github.com/LorenzoNorcini/Clickbait-Detector">[LorenzoNorcini] Clickbait-Detector</a>
- <a href="https://github.com/pfrcks/clickbait-detection">[pfrcks] clickbait-detection</a>
- <a href="https://github.com/saurabhmathur96/clickbait-detector">[saurabhmathur96] clickbait-detector</a>
- <a href="https://github.com/ventodiventu/clickbait-ml">[ventodiventu] clickbait-ml</a>

import os
import pandas as pd

## Functions

def merge_data(folder):
    merged_list = []
    
    for file in sorted(os.listdir(f'{folder}')):
        print(file)
    
        with open(f'{folder}/{file}', 'r') as f:
            for next_line in f.readlines():
                next_line = next_line.strip()
                if len(next_line) > 1:
                    merged_list.append(next_line)
                
    return merged_list

## Set input folder

data_in = '../data/00_raw/GitHub'

## Merge titles

clck_list = merge_data(f'{data_in}/_clickbait')

print(len(clck_list))

news_list = merge_data(f'{data_in}/_genuine')

print(len(news_list))

### [peterldowns] clickbait-classifier

files = ['buzzfeed', 'clickhole', 'dose', 'nytimes']

for file in files:
    df = pd.read_json(f'{data_in}/[peterldowns] clickbait-classifier/data/{file}.json')
    df['article_subject'] = df['article_title'].apply(lambda x: x.strip())

    news_new = list(df[df['clickbait'] == 0]['article_title'].str.strip())
    clck_new = list(df[df['clickbait'] == 1]['article_title'].str.strip())
    
    news_list += news_new
    clck_list += clck_new

    print(f'{file:15s} | Total: {df.shape[0]:>5d} | 0: {len(clck_new):>5d} | 1: {len(news_new):>5d}')

## Unique titles & Create DataFrames

clck_list = sorted(list(set(clck_list)))

clck_df = pd.DataFrame(clck_list, columns=['Title'])
display(clck_df)

news_list = sorted(list(set(news_list)))

news_df = pd.DataFrame(news_list, columns=['Title'])
display(news_df)

## Output data to csv

clck_df.to_csv(f'{data_in}/00_github_clck.csv', index=None, header=None)

news_df.to_csv(f'{data_in}/00_github_news.csv', index=None, header=None)
