==> notebooks/01_merge_titles_reddit.ipynb <==
# Reddit: Merge titles

- <a href="https://www.reddit.com/r/inthenews">/r/inthenews</a>
- <a href="https://www.reddit.com/r/news">/r/news</a>
- <a href="https://www.reddit.com/r/savedyouaclick">/r/savedyouaclick</a>

import numpy as np
import os
import pandas as pd

from pprint import pprint

## Set input folder

data_in = '../data/00_raw/pushshift/'

pprint(sorted(os.listdir(data_in)))

## Choose attributes of interest

aoi = ['created_utc',
       'num_comments',
       'score',
       'title']

## Clickbaits

### /r/savedyouaclick

df = pd.read_csv(f'{data_in}/savedyouaclick_all.csv',
                 index_col='Unnamed: 0')
df.head()

#### Split the titles from the answers

df['title'] = pd.Series([item[0].strip() if len(item[0].strip()) > 1 else np.nan
                         for item in df[aoi]['title'].str.split('|')])

#### Output data to csv format

df[aoi].to_csv(f'{data_in}/00_reddit_clck.csv')

## News

### Read data

#### /r/inthenews

df1 = pd.read_csv(f'{data_in}/inthenews_aoi.csv',
                  index_col='Unnamed: 0')
df1.head()

#### /r/news

df2 = pd.read_csv(f'{data_in}/news_gt1_aoi.csv',
                  index_col='Unnamed: 0')
df2.head()

### Merge data

df = pd.concat([df1, df2], sort=True, ignore_index=True)
df.head()

### Output data to csv format

df[aoi].to_csv(f'{data_in}/00_reddit_news.csv')
