==> notebooks/02_merge_by_class.ipynb <==
# Merge all titles

- Clickbait Challenge 2017


- GitHub
  - [bhargaviparanjape] clickbait
  - [LorenzoNorcini] Clickbait-Detector
  - [pfrcks] clickbait-detection
  - [saurabhmathur96] clickbait-detector
  - [ventodiventu] clickbait-ml


- Reddit
  - /r/inthenews
  - /r/news
  - /r/savedyouaclick

import numpy as np
import os
import pandas as pd

from pprint import pprint

data_in  = '../data/01_merged_by_source'
data_out = '../data/02_merged_by_class_raw'

pprint(sorted(os.listdir(data_in)))

## Clickbaits

#### Clickbait Challenge 2017

clck_cc2017= pd.read_csv(f'{data_in}/clck_cc2017.csv',
                          header=None)
clck_cc2017.columns = ['title']
clck_cc2017['src'] = ['cc2017' for _ in range(clck_cc2017.shape[0])]
clck_cc2017

#### GitHub

clck_github = pd.read_csv(f'{data_in}/clck_github.csv',
                          header=None)
clck_github.columns = ['title']
clck_github['src'] = ['github' for _ in range(clck_github.shape[0])]
clck_github

#### Reddit

clck_reddit = pd.read_csv(f'{data_in}/clck_reddit.csv',
                          index_col='Unnamed: 0')
clck_reddit = clck_reddit.dropna()
clck_reddit['src'] = ['reddit' for _ in range(clck_reddit.shape[0])]
clck_reddit.head()

### Merge all clickbaits

# clck = pd.concat([clck_cc2017[['title', 'src']],
#                   clck_github[['title', 'src']],
#                   clck_reddit[['title', 'src']]],
#                  ignore_index=True)

clck = pd.concat([clck_cc2017,
                  clck_github,
                  clck_reddit],
                 sort=True, ignore_index=True)

clck = clck.drop_duplicates(['title'])
clck = clck.sort_values(by='title')
clck.index = np.arange(1, clck.shape[0]+1)

display(clck.head())

print(f'There are {clck.shape[0]} unique clickbait titles.')

### Output data to csv format

clck.to_csv(f'{data_out}/clck_raw.csv')

<hr style="height: 5px" />

## News

#### Clickbait Challenge 2017

news_cc2017= pd.read_csv(f'{data_in}/news_cc2017.csv',
                          header=None)
news_cc2017.columns = ['title']
news_cc2017['src'] = ['cc2017' for _ in range(news_cc2017.shape[0])]
news_cc2017

#### GitHub

news_github = pd.read_csv(f'{data_in}/news_github.csv',
                          header=None)
news_github.columns = ['title']
news_github['src'] = ['github' for _ in range(news_github.shape[0])]
news_github

#### Reddit

news_reddit = pd.read_csv(f'{data_in}/news_reddit.csv',
                          index_col='Unnamed: 0')
news_reddit = news_reddit.dropna()
news_reddit['src'] = ['reddit' for _ in range(news_reddit.shape[0])]
news_reddit.head()

### Merge all news

# news = pd.concat([news_cc2017[['title', 'src']],
#                   news_github[['title', 'src']],
#                   news_reddit[['title', 'src']]],
#                  ignore_index=True)

news = pd.concat([news_cc2017,
                  news_github,
                  news_reddit],
                 sort=True, ignore_index=True)

news = news.drop_duplicates(['title'])
news = news.sort_values(by='title')
news.index = np.arange(1, news.shape[0]+1)

display(news.head())

print(f'There are {news.shape[0]} unique clickbait titles.')

### Output data to csv format

news.to_csv(f'{data_out}/news_raw.csv')
