==> notebooks/03_preprocessing_01_remove_non_letters_and_tag_numbers.ipynb <==
# Preprocessing: Remove non-letters and tag numbers

- Non-letters to " "
- Numbers to <number\>

import numpy as np
import os
import pandas as pd
import re

from pprint import pprint

## Functions

# %load '../snippets/count_chars.py'
def count_chars(series):
    charset = dict()
    
    for item in series:
        for char in set(item):
            if char not in charset.keys():
                charset[char] = 0
            charset[char] += 1
            
    print(len(charset), 'characters found.\n')
            
    return charset

## Set input/output folders

data_in = '../data/03_preprocessed_00'

pprint(sorted(os.listdir(data_in)))

data_out = '../data/03_preprocessed_01_non_letters_to_empty_space'

## Preprocessing

### Clickbait

clck = pd.read_csv(f'{data_in}/clck.csv', header=None)
clck.columns = ['headline']
clck

#### Remove punctuations...

pprint(count_chars(clck['headline']))

clck_non_letters = {
#     ' ': ' ',
    '!': ' ',
    '"': ' ',
    '#': ' ',
    '&': ' ',
    "'": ' ',
    '(': ' ',
    ')': ' ',
    '*': ' ',
    '+': ' ',
    ',': ' ',
    '-': ' ',
    '.': ' ',
    '/': ' ',
#     '0': ' ',
#     '1': ' ',
#     '2': ' ',
#     '3': ' ',
#     '4': ' ',
#     '5': ' ',
#     '6': ' ',
#     '7': ' ',
#     '8': ' ',
#     '9': ' ',
    ':': ' ',
    ';': ' ',
    '=': ' ',
    '?': ' ',
    '@': ' ',
#     'A': ' ',
#     'B': ' ',
#     'C': ' ',
#     'D': ' ',
#     'E': ' ',
#     'F': ' ',
#     'G': ' ',
#     'H': ' ',
#     'I': ' ',
#     'J': ' ',
#     'K': ' ',
#     'L': ' ',
#     'M': ' ',
#     'N': ' ',
#     'O': ' ',
#     'P': ' ',
#     'Q': ' ',
#     'R': ' ',
#     'S': ' ',
#     'T': ' ',
#     'U': ' ',
#     'V': ' ',
#     'W': ' ',
#     'X': ' ',
#     'Y': ' ',
#     'Z': ' ',
    '[': ' ',
    ']': ' ',
    '^': ' ',
#     'a': ' ',
#     'b': ' ',
#     'c': ' ',
#     'd': ' ',
#     'e': ' ',
#     'f': ' ',
#     'g': ' ',
#     'h': ' ',
#     'i': ' ',
#     'j': ' ',
#     'k': ' ',
#     'l': ' ',
#     'm': ' ',
#     'n': ' ',
#     'o': ' ',
#     'p': ' ',
#     'q': ' ',
#     'r': ' ',
#     's': ' ',
#     't': ' ',
#     'u': ' ',
#     'v': ' ',
#     'w': ' ',
#     'x': ' ',
#     'y': ' ',
#     'z': ' ',
    '|': ' ',
#     'Ü': ' ',
#     'à': ' ',
#     'á': ' ',
#     'â': ' ',
#     'ã': ' ',
#     'ä': ' ',
#     'å': ' ',
#     'ç': ' ',
#     'è': ' ',
#     'é': ' ',
#     'ê': ' ',
#     'ë': ' ',
#     'í': ' ',
#     'ñ': ' ',
#     'ó': ' ',
#     'ö': ' ',
#     'û': ' ',
#     'ü': ' ',
#     'ğ': ' '
}

clck_non_letters

for char in clck_non_letters.keys():
    clck['headline'] = clck['headline'].str.replace(char, 
                                                    clck_non_letters[char])
    
clck['headline']

### News

news = pd.read_csv(f'{data_in}/news.csv', header=None)
news.columns = ['headline']
news

pprint(count_chars(news['headline']))

news_non_letters = {
#     ' ': ' ',
    '!': ' ',
    '"': ' ',
    '#': ' ',
    '&': ' ',
    "'": ' ',
    '(': ' ',
    ')': ' ',
    '*': ' ',
    '+': ' ',
    ',': ' ',
    '-': ' ',
    '.': ' ',
    '/': ' ',
#     '0': ' ',
#     '1': ' ',
#     '2': ' ',
#     '3': ' ',
#     '4': ' ',
#     '5': ' ',
#     '6': ' ',
#     '7': ' ',
#     '8': ' ',
#     '9': ' ',
    ':': ' ',
    ';': ' ',
    '=': ' ',
    '?': ' ',
    '@': ' ',
#     'A': ' ',
#     'B': ' ',
#     'C': ' ',
#     'D': ' ',
#     'E': ' ',
#     'F': ' ',
#     'G': ' ',
#     'H': ' ',
#     'I': ' ',
#     'J': ' ',
#     'K': ' ',
#     'L': ' ',
#     'M': ' ',
#     'N': ' ',
#     'O': ' ',
#     'P': ' ',
#     'Q': ' ',
#     'R': ' ',
#     'S': ' ',
#     'T': ' ',
#     'U': ' ',
#     'V': ' ',
#     'W': ' ',
#     'X': ' ',
#     'Y': ' ',
#     'Z': ' ',
    '[': ' ',
    ']': ' ',
    '_': ' ',
#     'a': ' ',
#     'b': ' ',
#     'c': ' ',
#     'd': ' ',
#     'e': ' ',
#     'f': ' ',
#     'g': ' ',
#     'h': ' ',
#     'i': ' ',
#     'j': ' ',
#     'k': ' ',
#     'l': ' ',
#     'm': ' ',
#     'n': ' ',
#     'o': ' ',
#     'p': ' ',
#     'q': ' ',
#     'r': ' ',
#     's': ' ',
#     't': ' ',
#     'u': ' ',
#     'v': ' ',
#     'w': ' ',
#     'x': ' ',
#     'y': ' ',
#     'z': ' ',
    '|': ' ',
    '¡': ' ',
#     'Á': ' ',
#     'Å': ' ',
#     'É': ' ',
#     'Í': ' ',
#     'Î': ' ',
#     'Ü': ' ',
#     'ß': ' ',
#     'à': ' ',
#     'á': ' ',
#     'â': ' ',
#     'ã': ' ',
#     'ä': ' ',
#     'å': ' ',
#     'ç': ' ',
#     'è': ' ',
#     'é': ' ',
#     'ë': ' ',
#     'í': ' ',
#     'î': ' ',
#     'ñ': ' ',
#     'ó': ' ',
#     'ô': ' ',
#     'ö': ' ',
    '÷': ' ',
#     'ø': ' ',
#     'ú': ' ',
#     'ü': ' ',
#     'ā': ' ',
#     'ć': ' ',
#     'č': ' ',
#     'ę': ' ',
#     'ğ': ' ',
#     'ī': ' ',
#     'ł': ' ',
#     'ń': ' ',
#     'Ś': ' ',
#     'ś': ' ',
#     'Š': ' ',
#     'š': ' ',
#     'ū': ' ',
#     'ž': ' ',
#     'ș': ''    
}

news_non_letters

for char in news_non_letters.keys():
    news['headline'] = news['headline'].str.replace(char, 
                                                    news_non_letters[char])
news['headline']

## Tag numbers as <number\>

clck['headline'] = clck['headline'].str.replace(r'([0-9]+\s*[0-9]*)',
                                                ' <number> ')
clck['headline']

news['headline'] = news['headline'].str.replace(r'([0-9]+\s*[0-9]*)',
                                                ' <number> ')
news['headline']

## Output data to csv format

clck['headline'].to_csv(f'{data_out}/clck.csv',
                        header=None, index=None)

news['headline'].to_csv(f'{data_out}/news.csv',
                        header=None, index=None)
