==> notebooks/04_stats_sources.ipynb <==
# Statistics: Sources

import pandas as pd

## Set input folder

data_in = f'../data/03_preprocessed_00'

## Load data

clck_df = pd.read_csv(f'{data_in}/clck_full.csv')
news_df = pd.read_csv(f'{data_in}/news_full.csv')

## Source distribution analysis

### Number of entries

df = pd.DataFrame(index=['cc2017', 'github', 'reddit'],
                  columns=['clck', 'news'])
df = df.fillna(0)

for src in set(clck_df['src']):
    df.at[src, 'clck'] = len(clck_df[clck_df.src == src])
    
for src in set(news_df['src']):
    df.at[src, 'news'] = len(news_df[news_df.src == src])
    
df

df.sum(axis=0)

df.sum(axis=1)

### Percentage

df_pc = (df / df.sum().sum() * 100).round(2)
df_pc

df_pc.sum(axis=0)

df_pc.sum(axis=1)
